extracted ship model
